Moved profiling apps out to testing project
